scripts/requirements.txt
motor==3.4.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.4.0
pymongo==4.7.2